+++ morpheus-graphql-core/src/Data/Morpheus/Parser.hs
@@ -1,8 +1,8 @@
 {-# LANGUAGE NoImplicitPrelude #-}
 
 module Data.Morpheus.Parser
-  ( parseTypeDefinitions,
-    parseTypeSystemDefinition,
+  ( parseSchema,
+    parseTypeDefinitions,
     parseRequest,
     parseRequestWith,
   )
@@ -11,13 +11,16 @@ where
 import Data.ByteString.Lazy (ByteString)
 import Data.Morpheus.Ext.Result
   ( Eventless,
+    sortErrors,
   )
 import Data.Morpheus.Ext.SemigroupM ((<:>))
 import qualified Data.Morpheus.Parsing.Document.TypeSystem as P
   ( parseSchema,
     parseTypeDefinitions,
   )
-import Data.Morpheus.Parsing.Request.Parser (parseGQL)
+import Data.Morpheus.Parsing.Request.Parser
+  ( parseRequest,
+  )
 import Data.Morpheus.Schema.Schema (internalSchema)
 import Data.Morpheus.Types.IO
   ( GQLRequest (..),
@@ -25,7 +28,6 @@ import Data.Morpheus.Types.IO
 import Data.Morpheus.Types.Internal.AST
   ( ANY,
     CONST,
-    GQLQuery (..),
     Operation,
     Schema (..),
     TypeDefinition (..),
@@ -43,23 +45,22 @@ import Data.Morpheus.Validation.Query.Validation
   )
 import Relude hiding (ByteString)
 
-parseTypeSystemDefinition ::
+parseSchema ::
   ByteString -> Eventless (Schema VALID)
-parseTypeSystemDefinition =
-  P.parseSchema
-    >=> validateSchema
-      True
-      Config
-        { debug = False,
-          validationMode = FULL_VALIDATION
-        }
+parseSchema =
+  sortErrors
+    . ( P.parseSchema
+          >=> validateSchema
+            True
+            Config
+              { debug = False,
+                validationMode = FULL_VALIDATION
+              }
+      )
 
 parseTypeDefinitions ::
   ByteString -> Eventless [TypeDefinition ANY CONST]
 parseTypeDefinitions = P.parseTypeDefinitions
-
-parseRequest :: GQLRequest -> Eventless GQLQuery
-parseRequest = parseGQL
 
 parseRequestWith :: Config -> Schema VALID -> GQLRequest -> Eventless (Operation VALID)
 parseRequestWith config schema req = do